unified logging mechanisms in coreStar
everything is now in Config, and rely on the method that was previously
described in Debug. This has probably broken the effenciency of compile-time
flags where the compiler can optimise away logging that is not turned on. OTOH,
it saves the pain of manually turning the flags on and off and recompiling
every time. YMMV.

FIXME: the text in Debug is now slightly wrong. Also, Debug itself is not that
useful anymore...